removed old data dir
