--- data/README.md
@@ -1,5 +1,0 @@
-# Duet Data Design
-
-## Tech
-* Firebase instead (todo chat with Backend about how).
-* Dart as language to implement in.
